+++ lib/ansible/module_utils/remote_management/ucs.py
